refactor(PDFProView): :truck: PDFViewer moved to feature folder
